package update and type chagne

+++ bhive/scale.py
@@ -10,7 +10,7 @@ import argparse
 hx = HX711(5, 6)
 parser = argparse.ArgumentParser(description='beehive interface.')
 parser.add_argument('--reference_unit', type=float, help='The reference unit we divide the measurement by to get the desired weight.', default=20.544371)
-parser.add_argument('--offset', type=int, help='The offset in grams we substract from the measurement to tare it.', default=2115)
+parser.add_argument('--offset', type=float, help='The offset in grams we substract from the measurement to tare it.', default=2115)
 args = parser.parse_args()
 
 hx.set_reading_format("MSB", "MSB")
